src/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
mypy = "^1.6.1"
pydantic = "^2.4.2"
pytest = "^7.4.3"
coverage = "^7.3.2"
isort = "^5.12.0"
sentry-sdk = "^1.32.0"

[tool.poetry.scripts]
app = "app.main:hello"
sentry = "app.main:sentry_init"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
